+++ .github/workflows/CI.yml
@@ -25,11 +25,6 @@ jobs:
         with:
           version: ${{ matrix.version }}
           arch: ${{ matrix.arch }}
-      # GDPOptimizer is vendored and unregistered: dev it by path so
-      # Pkg.test can resolve it
-      - name: Develop the vendored GDPOptimizer.jl
-        run: julia --project=. -e 'using Pkg; Pkg.develop(path="GDPOptimizer.jl")'
-        shell: bash
       - uses: julia-actions/julia-runtest@latest
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v4
