+++ .github/workflows/Documentation.yml
@@ -14,9 +14,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@latest
       - name: Install dependencies
-        # dev the vendored (unregistered) GDPOptimizer together with the
-        # package so the docs project can resolve
-        run: julia --project=docs/ -e 'using Pkg; Pkg.develop([PackageSpec(path="GDPOptimizer.jl"), PackageSpec(path=pwd())]); Pkg.instantiate()'
+        run: julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
       - name: Build and deploy
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # For authentication with GitHub Actions token
