--- GDPOptimizer.jl/src/GDPOptimizer.jl
@@ -1,13 +1,0 @@
-module GDPOptimizer
-
-import MathOptInterface as MOI
-
-include("sets.jl")
-include("optimizer.jl")
-include("problem.jl")
-include("master.jl")
-include("nlp.jl")
-include("cuts.jl")
-include("loa.jl")
-
-end
