--- GDPOptimizer.jl/test/runtests.jl
@@ -1,9 +1,0 @@
-using Test
-import MathOptInterface as MOI
-using GDPOptimizer
-const GDPO = GDPOptimizer
-
-include("sets.jl")
-include("optimizer.jl")
-include("loa.jl")
-include("moi.jl")
