+++ test/runtests.jl
@@ -14,6 +14,7 @@ include("constraints/selector.jl")
 include("constraints/proposition.jl")
 include("constraints/disjunct.jl")
 include("constraints/indicator.jl")
+include("constraints/moidisjunction.jl")
 include("constraints/mbm.jl")
 include("constraints/bigm.jl")
 include("constraints/psplit.jl")
@@ -24,5 +25,4 @@ include("constraints/disjunction.jl")
 include("print.jl")
 include("solve.jl")
 include("extensions/InfiniteDisjunctiveProgramming.jl")
-include("extensions/GDPOptimizerDisjunctiveProgramming.jl")
